requirements.txt
streamlit

pillow

tensorflow==2.10.1
numpy==1.24.3
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.3.2
pandas==2.2.1
